lang_portal/backend_python/requirements.txt
# Core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# AWS
boto3==1.34.11
botocore==1.34.11

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Development
black==23.11.0
flake8==6.1.0
invoke==2.2.0 